Getting going with the OSX rendition of Stuff I Want from my Dotfiles. This one adds a pip module (for things like httpie and other python packages I reach for all the time). Tweaks to aliases and brew formulae, of course. Not tested at all yet.
One notable change: I've added code to lib/bash_profile to do a
conditional load (based on existance) of a file
HOME/.bash_local_only. This is for stuff like work config aliases and
stuff that just shouldn't be in public version control. Might eventually
submodule that file and put it in private VC.